MIC-8027: Move entities to separate files
