+++ TinkoffASDKUI/TinkoffASDKUI/Modules/AddNewCard/Assembly/IAddNewCardAssembly.swift
@@ -1,0 +1,20 @@
+//
+//  IAddNewCardAssembly.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 21.02.2023.
+//
+
+import Foundation
+
+protocol IAddNewCardAssembly {
+    func addNewCard(
+        customerKey: String,
+        output: IAddNewCardPresenterOutput?
+    ) -> AddNewCardViewController
+
+    func addNewCard(
+        customerKey: String,
+        onViewWasClosed: ((AddCardResult) -> Void)?
+    ) -> AddNewCardViewController
+}
