+++ TinkoffASDKUI/TinkoffASDKUI/Modules/AddNewCard/Models/AddNewCardSection.swift
@@ -1,0 +1,12 @@
+//
+//  AddNewCardSection.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 21.02.2023.
+//
+
+import Foundation
+
+enum AddNewCardSection {
+    case cardField
+}
