+++ TinkoffASDKUI/TinkoffASDKUI/Modules/AddNewCard/Views/ViewController/AddNewCardViewController.swift
@@ -21,44 +21,6 @@ import TinkoffASDKCore
 import UIKit
 import WebKit
 
-enum AddNewCardSection {
-    case cardField
-}
-
-// MARK: - AddNewCardOutput
-
-protocol IAddNewCardOutput: AnyObject {
-    /// Вызывается сразу при получении результата привязки карты на экране `AddNewCard`
-    ///
-    /// Потенциально может быть вызван несколько раз, поскольку при получении ошибки на экране `AddNewCard`
-    /// пользователь имеет возможность перезапустить процесс привязки карты.
-    /// По этой же причине вызов метода не всегда может означать скорое закрытие экрана `AddNewCard`
-    func addNewCardDidReceive(result: AddCardResult)
-
-    /// Вызывается единожды сразу после завершения анимации закрытия экрана
-    func addNewCardWasClosed(with result: AddCardResult)
-}
-
-extension IAddNewCardOutput {
-    func addNewCardDidReceive(result: AddCardResult) {}
-    func addNewCardWasClosed(with result: AddCardResult) {}
-}
-
-// MARK: - AddNewCardView
-
-protocol IAddNewCardView: AnyObject {
-    func reloadCollection(sections: [AddNewCardSection])
-    func showLoadingState()
-    func hideLoadingState()
-    func closeScreen()
-    func disableAddButton()
-    func enableAddButton()
-    func activateCardField()
-    func showOkNativeAlert(data: OkAlertData)
-}
-
-// MARK: - AddNewCardViewController
-
 final class AddNewCardViewController: UIViewController {
 
     // MARK: Dependencies
